+++ R/get.R
@@ -1,11 +1,20 @@
 
-#' Read configuration values.
+#' Read configuration values. Always use as `config::get()`.
 #'
 #' Read from the currently active configuration, retrieving either a
 #' single named value or all values as a list.
 #'
-#' For additional details see the package website at
-#'  \href{https://github.com/rstudio/config}{https://github.com/rstudio/config}.
+#' For additional details see <https://rstudio.github.io/config/>.
+#'
+#' @section Warning - Do not attach the package using library(config):
+#'
+#' We strongly recommend you use `config::get()` rather than attaching the
+#' package using `library(config)`.
+#'
+#' In fact, we strongly recommend you never use `library(config)`.
+#'
+#' The underlying reason is that the `get()` and `merge()` functions in
+#' `{config}` will mask these functions with the same names in base R.
 #'
 #' @param value Name of value (`NULL` to read all values)
 #'
@@ -27,6 +36,8 @@
 #' @return A list, or vector, corresponding to the contents of the config file.
 #'
 #' @seealso [is_active()], [merge()]
+#'
+#' @example inst/examples/example_get.R
 #'
 #' @export
 get <- function(value = NULL,
@@ -60,7 +71,10 @@ get <- function(value = NULL,
 
   # load the yaml
   config_yaml <- yaml::yaml.load_file(
-    file, eval.expr = FALSE, handlers = list(expr = function(x) parse(text = x))
+    file,
+    eval.expr = FALSE,
+    handlers = list(expr = function(x) parse(text = x)),
+    readLines.warn = FALSE
   )
 
   # get the default config (required)
@@ -96,22 +110,55 @@ get <- function(value = NULL,
 
   # check whether any expressions need to be evaluated recursively
 
-  eval_recursively <- function(x){
+  eval_issues <- list()
+  eval_env = new.env(parent = baseenv())
+  eval_fun <- function(expr, envir) {
+    tryCatch(
+      eval(expr, envir = envir),
+      error = function(e) {
+        eval_issues <<- append(
+          eval_issues,
+          paste(deparse(e$call), e$message, sep = ": ")
+        )
+        NULL
+      })
+  }
+  eval_recursively <- function(x, level = 1) {
     is_expr <- vapply(x, is.expression, logical(1))
-    x[is_expr] <- lapply(x[is_expr], eval, envir = baseenv())
-
     is_list <- vapply(x, is.list, logical(1))
-    x[is_list] <- lapply(x[is_list], eval_recursively)
 
+    if (level == 1) {
+      eval_env <- list2env(x[!is_expr & !is_list], envir = eval_env)
+    }
+    x[is_expr & !is_list] <- lapply(x[is_expr & !is_list], eval_fun, envir = eval_env)
+    x[is_list] <- lapply(x[is_list], eval_recursively, level = level + 1)
     x
   }
 
   active_config <- eval_recursively(active_config)
 
+  if (length(eval_issues)) {
+    msg <- paste("Attempt to assign nested list value from expression.",
+                 "Only directly assigned values can be used in expressions.",
+                 ngettext(length(eval_issues), "Original Error:\n",
+                          "Original Errors:\n"),
+                 sep = "\n")
+    stop(msg, paste("* ", eval_issues, collapse = "\n"), call. = TRUE)
+  }
+
   # return either the entire config or a requested value
   if (!is.null(value))
     active_config[[value]]
   else
-    structure(active_config, config = config, file = file)
+    structure(active_config, config = config, file = file, class = "config")
+
+}
+
+#' @export
+print.config <- function(x, ...) {
+  attr(x, "config") <- NULL
+  attr(x, "file") <- NULL
+  class(x) <- class(x)[-1]
+  NextMethod(x)
 }